Generalized RPS
Made new scripts for rock-paper-scissors-lizard-spock, rock-paper-scissors-surfer-unicorn-donut-fire, and a generalized version of RPS. The next steps will likely be making some kind of visualizer for the game using pygame. I'd like to make it easier to engineer these types of games in the future, and visualizing them is an important step for that to happen.